+++ packages/xchain-thorchain-amm/src/utils/index.ts
@@ -1,3 +1,2 @@
-export { getSingleSwap, getDoubleSwap } from './swap'
-export { getLiquidityProtectionData, getSlipOnLiquidity, getPoolShare, getLiquidityUnits } from './liquidity'
-export * from './midgard'
+export * from './evm-helper'
+export * from './eth-helper'